restructure with containers module
