+++ src/yarppg/__init__.py
@@ -25,18 +25,11 @@ __all__ = [
     "UiSettings",
 ]
 
+from .containers import Color, RegionOfInterest, RppgResult
 from .digital_filter import DigitalFilter
 from .helpers import bpm_from_frames_per_beat, frames_from_video, get_video_fps
 from .hr_calculator import HrCalculator, PeakBasedHrCalculator
 from .processors import ChromProcessor, FilteredProcessor, Processor
-from .roi import (
-    FaceMeshDetector,
-    RegionOfInterest,
-    RoiDetector,
-    SelfieDetector,
-    pixelate,
-    pixelate_mask,
-)
+from .roi import FaceMeshDetector, RoiDetector, SelfieDetector, pixelate, pixelate_mask
 from .rppg import Rppg
-from .rppg_result import Color, RppgResult
 from .settings import Settings, UiSettings, get_config
