+++ src/yarppg/roi/facemesh_segmenter.py
@@ -10,9 +10,10 @@ from mediapipe.tasks.python.components.containers import (
     landmark as landmark_module,  # type: ignore
 )
 
+from ..containers import RegionOfInterest
 from ..helpers import get_cached_resource_path
 from .detector import RoiDetector
-from .region_of_interest import RegionOfInterest, contour_to_mask
+from .region_of_interest import contour_to_mask
 
 MEDIAPIPE_MODELS_BASE = "https://storage.googleapis.com/mediapipe-models/"
 LANDMARKER_TASK = "face_landmarker/face_landmarker/float16/latest/face_landmarker.task"
